Fix PR #483 review issues: unused imports, calendar visibility, deduplication
- Remove unused `lte` imports from cron and generate routes (fixes lint/CI)
- Add `calendar` property to contextData schema type (fixes TS2769 build error)
- Extract shared `fetchCalendarContext` helper with visibility fix that excludes
  PRIVATE/ATTENDEES_ONLY events from other users in shared drives
- Apply same visibility fix to stats route (pulse/route.ts)
- Extract shared `PULSE_SYSTEM_PROMPT` to eliminate duplication across routes
